+++ src/composables/CodeEditor.ts
@@ -55,11 +55,13 @@ const completions = (context: any) => {
 export const useCodeEditor = (editorRef: Ref<HTMLElement | null>, {
   initialValue,
   emit,
-  commands
+  commands,
+  onPaste
 }: {
   initialValue: Ref<string>,
   emit?: any,
-  commands?: KeyBinding[]
+  commands?: KeyBinding[],
+  onPaste?: (data: string) => void
 }) => {
   const codeEditorStore = useCodeEditorStore()
 
@@ -79,6 +81,17 @@ export const useCodeEditor = (editorRef: Ref<HTMLElement | null>, {
       emit('update:modelValue', editorValue())
     })
 
+    const eventHandlers = []
+    if (onPaste) {
+      const handler = EditorView.domEventHandlers({
+        paste(event) {
+          const newValue = event.clipboardData?.getData('text')
+          if (newValue) onPaste(newValue)
+        }
+      })
+      eventHandlers.push(handler)
+    }
+
     const vimExtension = vim()
     codeMirrorEditor = new EditorView({
       extensions: [
@@ -88,6 +101,7 @@ export const useCodeEditor = (editorRef: Ref<HTMLElement | null>, {
         json(),
         autocompletion({ override: [completions] }),
         onChange,
+        eventHandlers,
         keymap.of([indentWithTab]),
         Prec.highest(keymap.of(commands || [])),
         keymap.of([{ key: 'Ctrl-Alt-l', mac: 'Ctrl-Cmd-l', run: beautifyEditorValue }]),